+++ src/main/java/usedbookshop/soobook/member/domain/Address.java
@@ -15,15 +15,15 @@ public class Address {
 
     private String area;
     private String roadName;
-    private Integer roadCode;
+    private Long roadCode;
 
-    private Address(String area, Integer roadCode, String roadName) {
+    private Address(String area, Long roadCode, String roadName) {
         this.area = area;
         this.roadName = roadName;
         this.roadCode = roadCode;
     }
 
-    public static Address createAddress(String area, Integer roadCode, String roadName) {
+    public static Address createAddress(String area, Long roadCode, String roadName) {
         return new Address(area, roadCode, roadName);
     }
 }
